Rename and move AuditInfo remote entity class
